requirements.txt
PySide6>=6.7.0
paddleocr>=3.4.0
paddlepaddle>=3.2.0
mss>=9.0.0
pywin32>=306
Pillow>=10.0.0
requests>=2.31.0
numpy>=2.0.0
